+++ genson/__init__.py
@@ -1,3 +1,3 @@
 from .schema.root import Genson
-from .schema.node import SchemaNode, InvalidSchemaError
-__all__ = ['Genson', 'SchemaNode', 'InvalidSchemaError']
+from .schema.node import SchemaNode, SchemaGenerationError
+__all__ = ['Genson', 'SchemaNode', 'SchemaGenerationError']